requirements.txt
# Cybersecurity Nexus - Python dependencies
# Pins are minimums; pip will pick the newest wheel that matches your Python.
# Recommended Python: 3.11 or 3.12 (best wheel coverage). 3.13/3.14 work too,
# just make sure pip is up to date so it can find the latest wheels.
neo4j>=5.20
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.10            # 2.10+ ships prebuilt wheels for Python 3.13/3.14
rich>=13.7
python-docx>=1.1          # for HIPAA SRA + Model Card docx output (optional fallback to .md)
